Add SidebarList component; modify sidebar filter functionality
Working on implementing a way for the sidebar filter on click to not
remove the recipe name from the sidebar list. Currently, it uses the
prevRecipes array, which works, but causes the issue of the recipe
getting removed if the user clicks to filter another recipe before
restoring the original one. If the regualar recipes array is used
then the recipe name disappears from both places until restored.